阵营与攻击大改&修复与其他模组的冲突

+++ src/main/java/com/hungteen/pvz/common/block/plants/LilyPadBlock.java
@@ -42,7 +42,7 @@ public class LilyPadBlock extends BushBlock {
 	public void entityInside(BlockState state, World worldIn, BlockPos pos, Entity entityIn) {
 		super.entityInside(state, worldIn, pos, entityIn);
 		if (worldIn instanceof ServerWorld && entityIn instanceof PVZZombieEntity ) {
-			if(entityIn.isAlive()) {
+			if(entityIn.isAlive() && ((PVZZombieEntity) entityIn).canBreakPlantBlock()) {
 			    worldIn.destroyBlock(new BlockPos(pos), true, entityIn);
 			}
 		}
